+++ docs/changes.rst
@@ -12,7 +12,7 @@ Not yet released.
 **Improvements**
 
 * Reduced memory usage for stastics.
-* :ref:`mt-deepl` performs better in :ref:`auto-translation`.
+* :ref:`mt-deepl` performs better in :ref:`auto-translation` and supports :ref:`glossary-mt`.
 
 **Bug fixes**